Feature: Admin Panel

+++ src/pages/MainPage/MainPageBookmarks/BookmarkCities/BookmarkCities.tsx
@@ -10,11 +10,11 @@ import {useNavigate} from "react-router-dom";
 const BookmarkCities = () => {
     const navigate = useNavigate();
     const [loading, setLoading] = useState(false);
-    const {dataCity, deleteCity, fetchData} = useFetchAdminCities();
+    const {dataCity, deleteCity, fetchDataCities} = useFetchAdminCities();
 
     useEffect(() => {
         setLoading(true);
-        fetchData();
+        fetchDataCities();
         console.log(dataCity);
         setLoading(false);
     }, [deleteCity]);
@@ -76,8 +76,12 @@ const BookmarkCities = () => {
                                 {
                                     <DropdownMenu
                                         deleteMethod={() => {
+                                            const assign = confirm("Видалити місто?")
+                                            if (assign) {
                                             deleteCity(item.id)
-                                            alert('Видалено')
+
+                                            }
+
                                         }}
                                         editMethod={() => handleCityClick(item.id)}
                                     />
